[rfc] Raise user-defined Zircon exception
This RFC introduces the `zx_thread_raise_exception` syscall, which
raises a user-defined Zircon exception. The first use case for this
syscall is for Starnix to signal to the debugger when one of its
processes calls `exec()`.  The debugger will use this signal to
determine whether the developer wants to attach to the process.

Change-Id: I8fdedf4677ff512e6b366b0f306fc10223f6499d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013525
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>