[lib][mock-mmio-reg] Fix MmioBuffer size calculation in MockMmioRegRegion
In https://fxrev.dev/896221 a typo was introduced that added rather than
multiplied the number of registers used by the MmioBuffer created using
MockMmioRegRegion::GetMmioBuffer(). In practice this has no effect on
the usage outside of buffer's get_size() returning the incorrect value
because all of the MmioOps of the crafted MmioBuffer are handled by the
MockMmioRegRegion's own vector and indexing, rather than a mapped VMO.
This CL fixes that math so get_size() should always match the intended
size of the register region.

Additionally, I added a quick non-functional cleanup in fake-mmio-reg
that I saw.

Fixed: 315010319
Change-Id: I635b817aba9ac66385f01a6001a81ad83fdf2de0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028861
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>