[wlan][policy] reformat new file for storage
The new file storage for saved networks is implemented but not
used yet. This CL changes the format of the new storage, and the new
storage is still unused after this change.

The previous format was a map of network identifiers to a list of
config data with that network identifier. That format was made to match
the previous way things were saved in stash and the way things are saved
in memory. For example, stash would write a value to a key, so the map
format made sense, but with the new file version the entire file is
rewritten each time.

Test: updated unit tests for the reformatted storage.
      (fx test wlan-stash-tests)
Change-Id: Ia48d5f340ee46e4168cfa0b4fb2a564504d3ef10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014465
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>