[serial][core] Add support for serialimpl FIDL
Make the serial core driver bind to both Banjo and FIDL serialimpl
drivers. Banjo support cannot be removed yet as it is still required by
ftdi-integration-test (ftdi will be moved to FIDL in a later change).
The core driver no longer implements the Banjo serial protocol as there
were no users of it left.

serialimpl read/write methods are called asynchronously, which will
eventually make the serial-async core driver and protocols redundant and
allow them to be removed. Also note that Write() is not called in a
loop, as the FIDL version of serialimpl guarantees that the entire
buffer is written.

Bug: 333094948
Multiply: serial-unittest
Test: ftdi-integration-test
Change-Id: I41d0a4fe20d8400b0fb4547c4032531f6bd65194
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031222
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>