[pkgdir] Refactor tests to work for CTF
A breaking change caused a problem with this test's use of Dictionary
methods, which means it will fail when run against the frozen version at
F19.

We stabilized the Namespace/Create method which can take in the
Dictionary and hide the implementation details, but unfortunately we do
not yet support extending namespaces with directories that are not
readonly. In this case, we need the directory to support EXECUTABLE
rights.

For now, we are changing this test to pass in a server_end for the
directory in its input realm arguments and return an empty dictionary
that is ignored. In the future we will refactor this test to support
getting its output from the dictionary directly.

Change-Id: I3856754cc3d63cd0511bda5e4659c01eb6c628a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034332
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>