[kernel][stats] Keep task runtime stats monotonic.
Address some of the issues in task runtime stat monitoring which can
cause non-monotonic behavior in stats like cpu runtime.

The main issue here comes down to where we actually observe our clock
to define when a state change has happened.  Prior to this change, we
were attempting to use a sequence lock to allow coherent reads of
thread stats without ever running the risk of blocking an updater.
The time of a state change, however, was being sampled outside of the
update transaction.  This allowed the following to potentially happen.

1) T1 (the updater thread) samples the clock at time B in order to
   record a transition from RUNNING to BLOCKED.  It has been in the
   RUNNING state since time A; A < B.
2) Before T1 starts the update transaction, thread T2 (the observer
   thread) arrives and successfully completes a read transaction,
   recording time C (C > B) as the time of the observation.  The
   observation contains:
2a) current_state       : RUNNING
2b) last_state_change   : A
2c) running_accumulator : X
3) T2 concludes that the total time in the running state at the time it
   made the observation is `X + (C - A)`
4) T1 performs its update transaction.  The state of the system is
   now:
4a) current_state       : BLOCKED
4b) last_state_change   : B
4c) running_accumulator : X + (B - A)
5) Finally, T2 queries the stats again.  This time it observes the
   total running time as `X + (B - A)`, but its prior observation was
   `X + (C - A)`, and we know that C > B, meaning it looks like this
   count has gone backwards.

The fix here comes from realizing that the current time, used to
update the last state change time or to compute the time in current
state, is effectively part of the payload being protected by sequence
lock.  From an accounting standpoint, the time at which a change takes
place *must* be defined as a time during the update transaction (when
read transactions cannot happen), and the time-in-current-state
computed during a read transaction *must* use an "observation time"
read during a successful read transaction.

To add a bit more unwanted complexity to all of this; clock reads are
not guaranteed to be bounded the memory order guarantees given by C++
or the current SeqLock implementation.  Clock registers are usually
_not_ memory and not subject to the same constraints.  So, when we do
observe the clock inside these update and read transactions, we need
to make sure to use the version of the kernel's clock read which
allows us to explicitly keep the actual observation of the clock
within the transaction as defined by the observation of (or
manipulation of) the SeqLock's sequence number.

These "fine-grained" runtime statistics are maintained per thread.
Queries of process stats currently involves holding the process
dispatcher's lock (so threads cannot enter or leave the process'
thread collection) and performing an O(n) calculation to sum the
totals across all threads in the process.

It _is_ possible to optimize this O(n) query away by applying a
similar technique used with the threads in order to constantly roll up
process-wide stats every time a thread in the process changes state.
This comes at a rather steep price, however.  In order to do this, we
need to make sure that the timestamps used for the thread update and
the process update have the same value.  This implies that the updater
is going to need to lock the process stats exclusively for the update
operation.  Prior to this, only the lock on the thread stats needed to
be held, which prevents concurrent reads, but never had to compete with
concurrent writes.  Holding a process level lock means that we could
have two threads competing to update at the same time, resulting in
contention where there wasn't any before.

Additionally, the math required to update the stats for a process is
slightly more complicated.  Whenever a thread exits state A, we need
to accumulate the total amount of time that threads were in state A
since the last recorded state change.  For thread stats, the count of
threads who were in state A before the state change is always 1; so if
X is the amount of time since the last state change, we just need to
add X to the accumulator.  For _process_ stats, however, things are
more complicated.  If the thread is exiting state A and entering state
B, we now have one less thread in A and one more in B.  In the case of
the thread stats, we new that the count of threads in A was 1 and the
count of threads in B was 0, but now these can both be different,
non-zero, numbers.  We need to multiply each count by X, and then sum
the products into the proper accumulators.  So, instead of one
subtraction and one addition, we are now paying for one subtraction,
two additions, and two multiplies.

This may not seem like much, but when coupled with the synchronization
requirements, it can actually become significant, to the point that it
starts to show significant regressions in context switch benchmarks
(where is where these state changes are almost always recorded).

So, for now, we skip the continuous roll-up of process specific stats,
and instead just record the basic thread level stats.  Context
switching and thread state changes are _extremely_ frequent, and small
incremental costs can add up extremely quickly.  In comparison,
querying of process stats happens extremely infrequently, so we let
the query operation take the O(n) performance hit (at least for now).

Bug: 42085921
Bug: 42085922
Change-Id: Ieef8ce375ebf59534874d5455376e45dab79add7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031275
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>