[rust] Update fuchsia-zircon-types for release
Removes a dependency on static-assertions in favor of a short custom
macro. Switches zerocopy to an optional dependency so it can be enabled
and disabled. Updates the crate manifest with relevant information and
bumps the version to a new alpha.

Change-Id: I873c5c7df6eea6019c638034113faafbdd4e6954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035092
Reviewed-by: Tyler Mandry <tmandry@google.com>
Fuchsia-Auto-Submit: David Koloski <dkoloski@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>