[aml-power] Remove dependency on GetDeviceInfo
In preparation to adding devicetree support removing the dependency on
GetDeviceInfo by using a generic constructor. The check for
availability of properties and protocols is already done as part of
device-enumeration tests and is rightly setup by the respective board driver.

Hence consolidated the power domain information into a single argument
for the ctor. Also removed some of the unused code related to luis.

For consistency of mapping, all platforms that have A53 cores are mapped
to little domain and A73 cores are mapped to big domain.

Bug: 333925701
Test: aml-meson-power-test, device enumeration tests
Change-Id: I83b3048f384054758dbbda0c475a8571307185e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027092
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>