[aml-sdmmc] Introduce no-Banjo driver build.
This code reorganization makes the AmlSdmmc driver class (almost)
purely DFv2+FIDL. Its derived class, AmlSdmmcWithBanjo, includes
the Banjo SDMMC protocol. No functional changes are intended in this
CL.

Banjo-related artifacts are purged from aml-sdmmc.h/cc and moved to
aml-sdmmc-with-banjo.h/cc. However in the BUILD.gn file, the original
"aml-sdmmc" name is kept for the driver build that includes Banjo.
This is to minimize the disruption to existing users of the driver.

Bug: b/328319349
Change-Id: I48d146921bf810398845d2ac8affe926b5a58c2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035377
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>