[ffx][package] Remove package plugin from ffx builtins
Removes ffx-package from being built in to the main ffx binary, allowing
it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx package` is invoked, but makes it so that both the tool and
ffx can build faster and with fewer resources making it easier to
iterate on them.

Bug: b/329326263
Change-Id: I9b324fb6eafd1037e9a27874e770db1e3cda6750
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916261
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>