[vsock-sshd-host] Introduce a small vsock variant of sshd
This variant is not meant to provide the same facilities as the typical
tcp port 22 listening variant. Instead it provides the minimal
environment necessary to allow an ffx connection to establish over the
vsock.

This variant is meant to be used in configuration where a netstack is
not available and we are running in an emulator which supports vsock.

Change-Id: Id5b7d10ee654a94f9353afa05410c7b7a057415b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030780
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Owners-Override: James Robinson <jamesr@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>