[fs_test] Add [[clang::lto_visibility_public]]
Currently, a bunch of memfs tests will fail on LTO builders when whole
program devirtualization is enabled. In particular, a bunch of memfs
tests fail because calls to the virtual functions Filesystem::Make and
FilesystemInstance::Mount were incorrectly being devirtualzied to
FilesystemImplWithDefaultMake::Make (in TestFilesystem::Create) and
BlobfsInstance::Mount (in TestFilesystem::Mount) for the memfs variant
of fs_tests. memfs provides their own overrides of these virtual
functions in memfs_fs_test.cc. These were incorrectly being devirted
because when we link the various tests in //src/storage/fs_test, the
definitions for the overriding classes in //src/storage/memfs are not
known in the LTO units for the fs_test executables. This means when
attempting to find all possible methods we can devirt to, the only
definitions seen are from //src/storage/fs_test.

The fix is to make the FilesystemInstance and FileSystem classes to have
public LTO visibility which indicates these classes are publicly
available outside the LTO linkage unit. This is the case here since
these classes are visible to other places outside the individual
//src/storage/fs_test binaries.

Bug: 42075686
Change-Id: I83274b24ad7778537541676236f1e53ffbc3888a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035252
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>