[power] Move domain creation to core driver
Power domains were created in the board driver using composite nodes.
There are a couple of issues with that mechanism: the logic will have to
be repeated in every board driver, this deviates from how gpio/i2c
children are created by the core drivers, this will be harder to create
when nested domain structure need to be created.

This CL updates the metadata passed to the power impl driver to include
domain information, which is forwarded to the power core. The power core
would internally create all the necessary domain nodes for relevant
power domain clients to bind to.
The motivation to make this change right now is it make the metadata and
node creation more streamlined for addition of devicetree support.

Bug: 333925701
Test: device enumeration tests, power-test
Change-Id: I2261e242f0f8124c8407e8a9055977090c80c3d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027106
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
API-Review: Christopher Anderson <cja@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>