[pkg-resolver] Do not cache blobfs results obtained outside the queue
Before adding the fetch of the meta.far blob to the queue,
pkg-resolver checks whether the meta.far is already resident. This
prevents blocking resolves of fully cached packages behind blob
fetches if the queue is already at its concurrency limit.

pkg-resolver was reusing the result of this check made outside the
queue in the fetch performed by the queue, which is incorrect,
because the queue may have already had a pending request to fetch
the blob, which would cause the result to be invalidated by the time
the queue started to process it, which would fail the fetch and
therefore the resolve.

With this change, pkg-resolver throws away the result and
rechecks for the blob's presence in the fetch performed by the queue.

Fixed: 336589151
Change-Id: I3722d82ea69a67d784032d0a1e4ab37f618e3085
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035239
API-Review: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: Sen Jiang <senj@google.com>