[package-directory] Prepare CTF tests for VmoFile
Update the package-directory CTF tests to accept the minor differences
VmoFile has with MetaFile, MetaAsFile, VmoBlob, and blobfs::Blob.

MetaFile and MetaAsFile both report to be writable in GetAttr requests.
They are not writable and VmoFile fixes this.

All of the files and directories are immutable so syncing them does
nothing. Returning either Ok or NOT_SUPPORTED is acceptable.

Blobfs blobs are only readable once they've been fully written but they
can be opened for reading before being fully written. The observer
event is used to indicate when a blobfs blob is readable. Fxblob blobs
can only be opened for reading once they've been written so the
observer event is not necessary. VmoBlob generated the event for
backwards compatibility. VmoFile doesn't generate the event.

It's unlikely that any client depends on these differences.

Bug: b/327633753
Change-Id: I800ea7a131411abc3bfd8f3665ea228e935f5c2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029312
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>