[netstack3] Move GMP to TimerContext2
As one of our older modules, GMP was more prone to internal races so we
had to shift some code around here. This is a good showcase of the
nudging power of the new TimerContext, which forces the timer handling
to happen under locks and exposes the previous races.

In this CL we tackle IGMP, MLD, and the generic GMP module all at once
since they're closely related. Note that there's still a possible race
in these modules when sending reports out of locks, that's left for a
future change.

Note that we've removed the assertions for extra groups on
enabling/disabling IPv6 on a device. Looking inside installed timers for
the MLD module is probably overkill there and the flow is already
covered by local tests.

Bug: 42083407
Change-Id: I07e8cd74e22c0b6944071d632858422415f6c71a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037936
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>