[debugger][analytics] Cleanse input for arbitrary strings
Certain strings that are reported to analytics will be verbatim as they
appeared in the console. Namely, invalid verbs and positional
arguments to valid verbs produce analytics with the text verbatim (in
the error message and argument reporting, respectively). These should be
cleansed before they are reported.

Change-Id: I8fe3ddfe66f7ed706e0506cf535df2962f98d181
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025729
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>