[display][coordinator] Rework C-style APIs for getting display info.
This CL replaces APIs in DisplayInfo and Controller that use
out-parameters and C-style strings with APIs that use C++ types. This
highlights the lifecycle and synchronization subtleties in the Client
code that uses the APIs. The intent is to make it easier to reason about
future changes to DisplayInfo representation and to its processing by
Controller and Client code.

While at it, this CL clarifies the DisplayInfo copy and move support.

Bug: 336394440
Test: fx test display-coordinator-unittest
Test: Booted core.qemu-x64 using ffx emu start --engine qemu
Change-Id: I7c776ac87ad82f0fcbf24d3ae55ccfe49e73e0c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037116
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>