[fidl] Optimize the client waker
On Fuchsia, we can read the channel when we are woken and then only open
the task that needs to be woken rather than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

This is a reland. I fixed the reason for the most recent revert in
https://fxrev.dev/1028132.

Change-Id: I18a43c197c01e230bfdaf46315b15b81dba0d91d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030892
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>