[fxfs] Extend profile timeout in delete_profile test.
Previously the profile was run for zero seconds before being explicitly
stopped. Checks prior the explicit stop were relying on the profile to
have not yet completed, so this was obviously a race.

Extended the timeout to 10 minutes, which is longer than the timeout
that infra will grant it anyways, so it must always wait for the
explicit stop.

Bug: b/327688777
Change-Id: I741fc3d328d6b9d949506646a3fd67d70d25ae8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039413
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>