[fake-display-stack] Replace CoordinatorDriver with non-driver Controller.
Previously in the fake-display-stack, the CoordinatorDriver was bound to
a mock device node. This change replaces the CoordinatorDriver with a
non-driver Controller instance; the binding between Controller and the
fake-dispaly engine is manually created.

This is part of the changes to migrate fake-display-stack out of the
driver framework v1.

BUg: 338140042
Change-Id: Id15378c0649a4219b3fd24b9a5ee99441d716d4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040474
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>