[cml] Replace merge macro with generic function
Replace the macro with a generic function. Instead of
calling this on every capability type, we can simply check that
the types match and that the names are the same.

This gives us better editor support for this function and
improve readability.

Change-Id: Ibc4841b1e2a3fc190733e03c5438c2916d44b8b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040232
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Gary Bressler <geb@google.com>